requirements.txt
bs4==0.0.1
kaggle==1.5.6
requests==2.24.0
snowflake-connector-python==2.2.8
